Exclude preview analytics and correct agent reference details
