+++ site/public/llms-full.txt
@@ -41,7 +41,7 @@ Either way the extension always starts and lists its tools, so if something is m
   - When: User works in Claude Code and wants meeting lifecycle skills and proactive hooks.
   - Install: `claude plugin marketplace add silverstein/minutes`
   - Best for: Best for workflow guidance, prep, debrief, and meeting coaching.
-  - Notes: 19 skills (brief, prep, debrief, graph, video review, and the rest of the meeting lifecycle) plus a meeting-analyst agent and SessionStart/PostToolUse hooks.
+  - Notes: Meeting lifecycle skills (brief, prep, debrief, graph, video review, and the rest of the meeting lifecycle) plus a meeting-analyst agent and SessionStart/PostToolUse hooks.
 - Pi coding agent
   - When: User works in Pi or wants Minutes to use pi as an opt-in local summarization agent.
   - Install: `pi`
@@ -51,7 +51,7 @@ Either way the extension always starts and lists its tools, so if something is m
   - When: User wants a menu bar app with one-click recording, Recall, and artifact drafting.
   - Install: `brew install --cask silverstein/tap/minutes`
   - Best for: Best for first recording, live capture, and post-meeting artifact work.
-  - Notes: Tauri v2, macOS only today. Includes command palette, dictation hotkey, and live mode toggle. Updates remain manual until a hosted manifest and rollback UX exist.
+  - Notes: Tauri v2 desktop app for macOS, with an unsigned Windows preview installer. Calendar suggestions, call detection, and the native dictation hotkey are macOS-only. Updates remain manual until a hosted manifest and rollback UX exist.
 
 ## Canonical entry points
